requirements.txt
# Core RAG dependencies
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langgraph>=0.1.0

# Vector store and embeddings
faiss-cpu>=1.8.0
sentence-transformers>=2.3.1

# PDF processing
pypdf>=4.0.1

# LLM integration
groq>=0.4.2
requests>=2.31.0

# Evaluation
ragas>=0.1.1
datasets>=2.16.1
pandas>=2.0.0

# Web interface
streamlit>=1.31.1

# Utilities
python-dotenv>=1.0.0
psutil>=5.9.0
numpy>=1.24.0

# Development 
pytest>=8.0.0
black>=24.0.0
flake8>=7.0.0